fix(security): consolidate HTTP timeout accessors and enhance timeout handling
